Backend/requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.8
cloudinary==1.42.1
distlib==0.3.9
dj-database-url==2.3.0
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
filelock==3.16.1
gunicorn==23.0.0
h11==0.14.0
idna==3.10
mpmath==1.3.0
packaging==24.2
pillow==10.4.0
platformdirs==4.3.6
psycopg2-binary==2.9.10
pycparser==2.22
python-decouple==3.8
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
sympy==1.13.3
typing_extensions==4.12.2
Unidecode==1.3.8
urllib3==2.2.3
uvicorn==0.34.0
virtualenv==20.28.0
whitenoise==6.8.2
